Use static list of server A records when forming cluster
This is the first step to be able to support TLS by using a wildcard
certificate between the routes.  Instead of using auto discovery, the
cluster will be formed by using using the full list of hosts that will
be forming the cluster, and since it is an explicit list they will
continue to reconnect until the cluster is formed.

Note that this approach requires NATS image v1.0.4 in order to
wait for the A records from all the involved pods to be ready
which could take a bit occasionally.