Allow Pod Template to provide initContainers, Volumes, and Containers
Previously, `NewNatsPodSpec` would initialize it the spec from a deep
copy of the PodTemplate, if it was provided by the user.  However, it then
overwrites initContainers, Volumes, and Containers with item lists that
are build at runtime.

This commit changes the behavior to append to the existing lists, enabling
InitContainers, Volumes, and Containers to be specified in the pod template.

Additionally, this adds a `VolumeMounts` item to the PodPolicy so that
volumes provided in the template, can be mounted into the nats container.