pkg/util: make leafnodeconfig independent
Currently, the leafnode configuration is only added to the configuration
secret if a gateway configuration is specified. However, these are two
orthogonal parts of the NATS configuration: one should be able to
specify a leafnode configuration without specifying a gateway
configuration, e.g. as shown here:
https://github.com/nats-io/k8s/blob/master/nats-server/simple-nats.yml#L23-L25.

This commit separates the two configurations.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>